requirements.txt
Django~=5.1
djangorestframework~=3.15.2
python-dotenv~=1.0.1
drf-yasg==1.21.7
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
smsaero_api==3.0.0
django-filter==24.3

